requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-registration==2.4.1
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
whitenoise==4.1
wrapt==1.10.11
